requirements.txt
annotated-types==0.7.0
Authlib==1.6.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dimod==0.12.17
diskcache==5.6.3
dwave-cloud-client==0.13.0
dwave-optimization==0.3.0
dwave-preprocessing==0.6.6
dwave-samplers==1.3.0
dwave-system==1.26.0
dwave_networkx==0.8.15
fasteners==0.19
homebase==1.0.1
idna==3.10
importlib_metadata==8.5.0
MarkupSafe==2.1.5
minorminer==0.2.15
networkx==3.3
numpy==2.1.1
orjson==3.10.7
packaging==24.1
plucky==0.4.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.14.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
joblib==1.3.2
matplotlib==3.8.1
